Check component is mounted before async setState
I got a warning while using this library. Check whether the component is mounted before we callback with a setState call. This 

```Warning: Can't call setState (or forceUpdate) on an unmounted component. This is a no-op, but it indicates a memory leak in your application. To fix, cancel all subscriptions and asynchronous tasks in the componentWillUnmount method.
    in SimpleImg (created by Context.Consumer)```